+++ src/components/BrowserWindow/OnlyofficeEditor.tsx
@@ -12,6 +12,11 @@ interface OnlyOfficeEditorProps {
   config?: code;
   isDemo?: boolean;
   isForm?: boolean;
+  externalScript?: {
+    beforeDocumentReady: string;
+    onDocumentReady: string;
+    otherFunctional: string;
+  };
 }
 
 async function createJWT(json: object, secret: string): Promise<string | null> {
@@ -81,11 +86,11 @@ const getDocumentName = (isDemo: boolean = false, isForm: boolean = false): stri
   return isDemo ? (isForm ? "demo-invoice" : "demo") : "new";
 };
 
-const createDocumentConfig = (fileType: string, templateUrl: string, isDemo: boolean = false, isForm: boolean = false): object => {
+const createDocumentConfig = (fileType: string, templateUrl: string, title?: string, isDemo: boolean = false, isForm: boolean = false): object => {
   return {
     fileType,
     key: `doc-${Date.now()}`,
-    title: `Example Document.${fileType}`,
+    title: title || `Example Document.${fileType}`,
     url: templateUrl ? templateUrl : `https://static.onlyoffice.com/assets/docs/samples/${getDocumentName(isDemo, isForm)}.${fileType}`
   };
 };
@@ -111,15 +116,22 @@ function deepMergePreferFirst(a: any, b: any): any {
   return result;
 }
 
+const scriptId: string = "editorScript";
+
 const addScript = async (server: string, secret: string, fileType: string, code: string, theme: string, templateUrl: string, zoom: number,
-                         externalConfig?: code, isDemo: boolean = false, isForm: boolean = false): Promise<void> => {
+                         externalConfig?: code, externalScript?: OnlyOfficeEditorProps["externalScript"], isDemo: boolean = false, isForm: boolean = false): Promise<void> => {
   const scriptConfig = document.createElement("script");
+  scriptConfig.id = scriptId;
   scriptConfig.type = "text/javascript";
 
-  const documentConfig = createDocumentConfig(fileType, templateUrl, isDemo, isForm);
+  const documentConfig = createDocumentConfig(fileType, templateUrl, externalConfig?.document?.title, isDemo, isForm);
 
   if (externalConfig?.editorConfig?.customization?.logo?.image) {
     externalConfig.editorConfig.customization.logo.image = new URL("/assets/images/try-docs/example-logo.png", window.location.origin).href;
+  }
+
+  if (externalConfig?.editorConfig?.customization?.customer?.logo) {
+    externalConfig.editorConfig.customization.customer.logo = new URL("/assets/images/try-docs/example-logo-big.png", window.location.origin).href;
   }
 
   const config = deepMergePreferFirst(
@@ -147,12 +159,14 @@ const addScript = async (server: string, secret: string, fileType: string, code:
       window.docEditor = null;
     }
 
+    ${externalScript?.beforeDocumentReady}
     window.onDocumentReady = function() {
       window.connector = docEditor.createConnector();
       connector.callCommand(() => {
         ${code}
       }, () => {
       });
+      ${externalScript?.onDocumentReady}
     };
 
     var config = ${JSON.stringify(config)};
@@ -162,6 +176,7 @@ const addScript = async (server: string, secret: string, fileType: string, code:
     };
 
     window.docEditor = new DocsAPI.DocEditor("placeholder", config);
+    ${externalScript?.otherFunctional}
   `;
 
   document.body.appendChild(scriptConfig);
@@ -175,7 +190,12 @@ const OnlyOfficeEditor: React.FC<OnlyOfficeEditorProps> = ({
   zoom,
   config,
   isDemo = false,
-  isForm = false
+  isForm = false,
+  externalScript = { 
+    beforeDocumentReady:"",
+    onDocumentReady:"",
+    otherFunctional:"" 
+  }
 }) => {
   const {
     siteConfig: {customFields},
@@ -198,17 +218,18 @@ const OnlyOfficeEditor: React.FC<OnlyOfficeEditorProps> = ({
         };
         scriptApi.onload = () => {
           document.documentElement.setAttribute("data-script-api-state", "2");
-          addScript(documentServer, documentServerSecret, fileType, code, colorMode, templateUrl, zoom, config, isDemo, isForm);
+          addScript(documentServer, documentServerSecret, fileType, code, colorMode, templateUrl, zoom, config, externalScript, isDemo, isForm);
         };
 
         document.documentElement.setAttribute("data-script-api-state", "1");
         document.body.appendChild(scriptApi);
       } else {
-        addScript(documentServer, documentServerSecret, fileType, code, colorMode, templateUrl, zoom, config, isDemo, isForm);
+        addScript(documentServer, documentServerSecret, fileType, code, colorMode, templateUrl, zoom, config, externalScript, isDemo, isForm);
       }
     }
 
     return () => {
+      document.getElementById(scriptId).remove();
     };
   }, []);