+++ src/components/BrowserWindow/index.ts
@@ -1,0 +1,5 @@
+export { default as OnlyofficeEditor } from "./OnlyofficeEditor";
+
+export { default as FormExternalToolbar } from "./ExternalToolbars/FormExternalToolbar/FormExternalToolbar";
+export { default as CommentsExternalToolbar } from "./ExternalToolbars/CommentsExternalToolbar/CommentsExternalToolbar";
+export { default as ReviewChangesExternalToolbar } from "./ExternalToolbars/ReviewChangesExternalToolbar/ReviewChangesExternalToolbar";